hugging/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
packaging==24.2
pillow==11.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
scapy==2.6.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
